Handled Delivery Addresses and Media Upload

+++ app/Http/Controllers/Api/AddressController.php
@@ -17,9 +17,12 @@ class AddressController extends Controller{
         $user = $this->user();
         $unique_id = $this->createUniqueId('addresses');
 
+        $default = $user->addresses()->count() > 0 ? $this->no : $this->yes;
+
         Address::create($request->safe()->merge([
             'user_id' => $user->unique_id,
-            'unique_id' => $unique_id
+            'unique_id' => $unique_id,
+            'default' => $default
         ])->all());
 
         return $this->returnMessageTemplate(true, "Address Created", [
@@ -36,11 +39,14 @@ class AddressController extends Controller{
         ]);
     }
 
-    function update(UpdateAddressRequest $request, Address $address){
+    function update(UpdateAddressRequest $request, $id){
+        $address = Address::find($id);
+        if(!$address) return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_found', "Address"));
+
         $user = $this->user();
 
         if($address->user_id !== $user->unique_id)
-                    return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_owner', "Address", 'the current User'));
+        return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_owner', "Address", 'the current User'));
 
         $address->update($request->safe()->all());
 
@@ -50,15 +56,48 @@ class AddressController extends Controller{
         ]);
     }
 
-    function delete(Address $address){
+    function single($id){
+        $address = Address::find($id);
+        if(!$address) return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_found', "Address"));
+        return $this->returnMessageTemplate(true, "", [
+            'address' => $address
+        ]);
+    }
+
+    function delete($id){
+        $address = Address::find($id);
+
+        if(!$address) return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_found', 'Address'));
         $user = $this->user();
+
         if($address->user_id !== $user->unique_id)
-                        return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_owner', "Address". 'the current User'));
+            return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_owner', "Address", 'the current User'));
+
         $address->delete();
 
         return $this->returnMessageTemplate(true, $this->returnSuccessMessage('deleted', 'Address'), [
             'user' => $user->with(['addresses']),
             'addresses' => $user->addresses
+        ]);
+    }
+
+    function setDefault($id){
+        $address = Address::find($id);
+        if(!$address) return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_found', "Address"));
+
+        $user = $this->user();
+
+        if($address->user_id !== $user->unique_id)
+                        return $this->returnMessageTemplate(false, $this->returnErrorMessage('not_owner', "Address", 'the current user'));
+
+        $user->addresses()->where('default', $this->yes)->update(['default' => $this->no]);
+
+        $address->default = $this->yes;
+        $address->save();
+
+        return $this->returnMessageTemplate(true, "$address->name address has been set to default", [
+            'addresses' => $user->addresses,
+            'address' => $address
         ]);
     }
 }
